Refactor CodeStatus into own file
