+++ astoria/common/messages/astprocd.py
@@ -1,20 +1,10 @@
 """Message schemas for astprocd."""
-from enum import Enum
 from typing import Optional
 
+from astoria.common.code_status import CodeStatus
 from astoria.common.ipc import ManagerMessage
 
 from .astdiskd import DiskInfo
-
-
-class CodeStatus(str, Enum):
-    """Status of the running code."""
-
-    STARTING = "code_starting"
-    RUNNING = "code_running"
-    KILLED = "code_killed"
-    FINISHED = "code_finished"
-    CRASHED = "code_crashed"
 
 
 class ProcessManagerMessage(ManagerMessage):
