+++ tests/common/messages/test_astprocd.py
@@ -2,18 +2,9 @@
 from pathlib import Path
 
 from astoria import __version__
+from astoria.common.code_status import CodeStatus
 from astoria.common.disks import DiskInfo, DiskType, DiskUUID
-from astoria.common.messages.astprocd import CodeStatus, ProcessManagerMessage
-
-
-def test_code_status_enum() -> None:
-    """Test that CodeStatus has the right values."""
-    assert CodeStatus.STARTING.value == "code_starting"
-    assert CodeStatus.RUNNING.value == "code_running"
-    assert CodeStatus.KILLED.value == "code_killed"
-    assert CodeStatus.FINISHED.value == "code_finished"
-    assert CodeStatus.CRASHED.value == "code_crashed"
-    assert len(CodeStatus) == 5
+from astoria.common.messages.astprocd import ProcessManagerMessage
 
 
 def test_proc_manager_fields() -> None:
